# docs/swagger.yaml
definitions:
  data.Author:
    properties:
      id:
        type: integer
      name:
        type: string
    type: object
  data.Book:
    properties:
      authors:
        items:
          $ref: '#/definitions/data.Author'
        type: array
      edition:
        type: integer
      id:
        type: integer
      name:
        type: string
      publication_year:
        type: integer
    type: object
info:
  contact: {}
paths:
  /author:
    get:
      consumes:
      - application/json
      description: list all authors
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/data.Author'
            type: array
      summary: Author
  /book:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/data.Book'
            type: array
      summary: Book
    post:
      consumes:
      - application/json
      parameters:
      - description: Book JSON
        in: body
        name: book
        required: true
        schema:
          $ref: '#/definitions/data.Book'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/data.Book'
        "400":
          description: Bad Request
      summary: Book
  /book/{id}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: Book ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: Delete Book
    get:
      consumes:
      - application/json
      parameters:
      - description: Book ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/data.Book'
        "404":
          description: Not Found
      summary: Book
    put:
      consumes:
      - application/json
      parameters:
      - description: Book JSON
        in: body
        name: book
        required: true
        schema:
          $ref: '#/definitions/data.Book'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/data.Book'
        "400":
          description: Bad Request
      summary: Book
swagger: "2.0"